requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
contourpy==1.0.5
cycler==0.11.0
docker-pycreds==0.4.0
fonttools==4.37.4
gitdb==4.0.9
GitPython==3.1.29
gym==0.21.0
gym-notices==0.0.8
h5py==3.7.0
Hydra==2.5
idna==3.4
importlib-metadata==4.13.0
kiwisolver==1.4.4
matplotlib==3.6.1
networkx==2.8.7
numpy==1.23.4
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
pathtools==0.1.2
Pillow==9.2.0
pkg_resources==0.0.0
promise==2.3
protobuf==4.21.7
psutil==5.9.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
sentry-sdk==1.9.10
setproctitle==1.3.2
Shapely==1.8.5.post1
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
stable-baselines3==1.6.2
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
typing_extensions==4.4.0
urllib3==1.26.12
wandb==0.13.4
zipp==3.9.0
